requirements.txt
python-dotenv
requests
selenium
bs4
webdriver_manager